finish ver1

+++ server/src/main/java/com/se/ssps/server/entity/configuration/MaxFileSize.java
@@ -1,10 +1,12 @@
 package com.se.ssps.server.entity.configuration;
 
+// import jakarta.persistence.Entity;
+// import jakarta.persistence.Table;
 import lombok.*;
 
 @Setter
 @Getter
 @AllArgsConstructor
 public class MaxFileSize {
-    private Integer maxFileSize;
+    private double maxFileSize;
 }
